+++ rust-plugins/sass/src/lib.rs
@@ -113,13 +113,17 @@ impl Plugin for FarmPluginSass {
     _context: &std::sync::Arc<farmfe_core::context::CompilationContext>,
     _hook_context: &farmfe_core::plugin::PluginHookContext,
   ) -> farmfe_core::error::Result<Option<farmfe_core::plugin::PluginLoadHookResult>> {
-    if self.regex.is_match(param.resolved_path) {
-      let content = fs::read_file_utf8(param.resolved_path).unwrap();
-      return Ok(Some(farmfe_core::plugin::PluginLoadHookResult {
-        content,
-        module_type: ModuleType::Custom(String::from("sass")),
-      }));
+    if param.query.is_empty() && self.regex.is_match(param.resolved_path) {
+      let content = fs::read_file_utf8(param.resolved_path);
+
+      if let Ok(content) = content {
+        return Ok(Some(farmfe_core::plugin::PluginLoadHookResult {
+          content,
+          module_type: ModuleType::Custom(String::from("sass")),
+        }));
+      }
     }
+
     Ok(None)
   }
 
@@ -161,7 +165,7 @@ impl Plugin for FarmPluginSass {
         .common
         .importers
         .push(sass_embedded::SassImporter::FileImporter(import_collection));
-      string_options.url = None;
+      string_options.url = Some(Url::from_file_path(param.resolved_path).unwrap());
 
       let compile_result = sass.compile_string(&content, string_options).map_err(|e| {
         farmfe_core::error::CompilationError::TransformError {