+++ js-plugins/vue/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @farmfe/js-plugin-vue
 
+## 3.1.0
+
+### Minor Changes
+
+- Bump version
+
+## 3.0.1
+
+### Patch Changes
+
+- a569977: Update filters to support resolved paths with query
+- Updated dependencies [a569977]
+- Updated dependencies [a569977]
+  - @farmfe/core@0.13.19
+
 ## 3.0.0
 
 ### Patch Changes